Select the preferred maze generation method.
The "Twisted Halls With Rooms" method has support, but is not implemented, and is never used if selected - instead defaulting to the "Randomness with limits" method. The "Pure randomness" method, on the other hand, is fully implemented and will work if chosen. Just be warned... this method is highly likely to make unsolvable mazes!